requirements.txt
psycopg2
requests
Scrapy
python-dateutil
